Move config_cache() method from api.py to config_cache.py
