+++ dbcollection/core/config_cache.py
@@ -10,6 +10,65 @@ import shutil
 from dbcollection.core.cache import CacheManager
 
 
+def config_cache(field=None, value=None, delete_cache=False, delete_cache_dir=False,
+                 delete_cache_file=False, reset_cache=False, verbose=True, is_test=False):
+    """Configure the cache file.
+
+    This method allows to configure the cache file directly by selecting
+    any data field/value. The user can also manually configure the file
+    if he/she desires.
+
+    To modify any entry in the cache file, simply input the field name
+    you want to change along with the new data you want to insert. This
+    applies to any field/data in the file.
+
+    Another thing available is to reset/clear the entire cache paths/configs
+    from the file by simply enabling the 'reset_cache' flag to true.
+
+    Also, there is an option to completely remove the cache file+folder
+    from the disk by enabling 'delete_cache' to True. This will remove the
+    cache dbcollection.json and the dbcollection/ folder from disk.
+
+    Parameters
+    ----------
+    field : str, optional
+        Name of the field to update/modify in the cache file.
+    value : str, list, tuple, optional
+        Value to update the field.
+    delete_cache : bool, optional
+        Delete/remove the dbcollection cache file + directory.
+    delete_cache_dir : bool, optional
+        Delete/remove the dbcollection cache directory.
+    delete_cache_file : bool, optional
+        Delete/remove the dbcollection.json cache file.
+    reset_cache : bool, optional
+        Reset the cache file.
+    verbose : bool, optional
+        Displays text information (if true).
+    is_test : bool, optional
+        Flag used for tests.
+
+    Examples
+    --------
+    Delete the cache by removing the dbcollection.json cache file.
+    This will NOT remove the file contents in dbcollection/. For that,
+    you must set the *delete_cache_dir* argument to True.
+
+    >>> import dbcollection as dbc
+    >>> dbc.config_cache(delete_cache_file=True)
+    """
+    manager = ConfigAPI(field=field,
+                        value=value,
+                        delete_cache=delete_cache,
+                        delete_cache_dir=delete_cache_dir,
+                        delete_cache_file=delete_cache_file,
+                        reset_cache=reset_cache,
+                        verbose=verbose,
+                        is_test=is_test)
+
+    manager.run()
+
+
 class ConfigAPI(object):
     """Cache configuration API class.
